requirements.txt
tkinter
threading
imaplib
email
python-dotenv
requests
